trip manager update All requested changes made
