+++ core/src/main/java/com/bankbazaar/core/manager/TripActivityManager.java
@@ -1,6 +1,5 @@
 package com.bankbazaar.core.manager;
 
-import com.bankbazaar.core.model.TripActivityCompositeKey;
 import com.bankbazaar.core.model.TripActivityEntity;
 import com.bankbazaar.core.model.TripEntity;
 import com.bankbazaar.core.repository.TripActivityRepository;
@@ -13,7 +12,7 @@ import java.util.List;
 import java.util.Optional;
 @Slf4j
 @Service
-public class TripActivityEntityManager {
+public class TripActivityManager {
     @Autowired
     private TripActivityRepository tripActivityRepository;
 
@@ -26,8 +25,8 @@ public class TripActivityEntityManager {
      */
     public TripActivityEntity saveTripActivity(TripActivityEntity data)
     {
-        Optional<TripEntity> tripId = tripRepository.findById(data.getTripId().getTripId());
-        data.setTripId(tripId.get());
+        Optional<TripEntity> trip = tripRepository.findById(data.getTripId().getTripId());
+        data.setTripId(trip.get());
         return tripActivityRepository.save(data);
     }
 
@@ -37,7 +36,7 @@ public class TripActivityEntityManager {
      */
     public TripActivityEntity updateTripActivity(TripActivityEntity data)
     {
-        Optional<TripActivityEntity> presentData = exists(data.getTripId().getTripId(),data.getActivityId());
+        Optional<TripActivityEntity> presentData = exists(data.getActivityId());
         if(presentData.isPresent())
         {
             TripActivityEntity newData = updateData(presentData.get(), data);
@@ -46,12 +45,6 @@ public class TripActivityEntityManager {
         }
         return null;
     }
-    /**
-     * Get record by ID
-     */
-    public Optional<TripActivityEntity> getTripActivityById(Long tripId, Long activityId) {
-        return tripActivityRepository.findById(new TripActivityCompositeKey(tripId,activityId));
-    }
 
     /**
      * get all activities using tripId
@@ -59,12 +52,12 @@ public class TripActivityEntityManager {
      */
     public List<TripActivityEntity> getActivityTripId(Long tripId) {
 
-        return tripActivityRepository.getByTripId(tripId);
+        return tripActivityRepository.findAllByTripIdTripId(tripId);
     }
 
-    private Optional<TripActivityEntity> exists(Long tripId, Long activityId)
+    private Optional<TripActivityEntity> exists(Long activityId)
     {
-        return tripActivityRepository.findById(new TripActivityCompositeKey(tripId,activityId));
+        return tripActivityRepository.findById(activityId);
     }
     private TripActivityEntity updateData(TripActivityEntity presentData, TripActivityEntity tripActivity)
     {
