--- core/src/main/java/com/bankbazaar/core/model/TripActivityCompositeKey.java
@@ -1,16 +1,0 @@
-package com.bankbazaar.core.model;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class TripActivityCompositeKey implements Serializable {
-    private Long tripId;
-    private Long activityId;
-}
