+++ core/src/main/java/com/bankbazaar/core/repository/TripUserMapRepository.java
@@ -3,14 +3,10 @@ package com.bankbazaar.core.repository;
 import com.bankbazaar.core.model.TripUserMapEntity;
 import com.bankbazaar.core.model.TripUserCompositeKey;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
 public interface TripUserMapRepository extends JpaRepository<TripUserMapEntity, TripUserCompositeKey> {
-    @Query("SELECT t FROM TripUserMapEntity t WHERE t.userId.userId = ?1")
-    public List<TripUserMapEntity> getByUserId(Long userId);
-
-    @Query("SELECT t FROM TripUserMapEntity t WHERE t.tripId.tripId = ?1")
-    public List<TripUserMapEntity> getByTripId(Long tripId);
+    public List<TripUserMapEntity> findAllByUserIdUserId(Long userId);
+    public List<TripUserMapEntity> findAllByTripIdTripId(Long tripId);
 }
