+++ core/src/main/java/com/bankbazaar/core/repository/UserRepository.java
@@ -2,7 +2,6 @@ package com.bankbazaar.core.repository;
 
 import com.bankbazaar.core.model.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
